Fixed up glade file scrollbar policy. Hopefully they won't suddenly always appear again.
Modified pulseblaster variable names to be more consistent with other classes

Modified the name/channel passed to AO/DO objects within a DDS object to better reflect what they are (in both the novatech and the pulseblaster)

Redesigned the save front panel state system to be within a separate object. This is only is use via the file->save option, and is currently untested.

Saving and loading at close and startup respectively still uses the old system and should not have changed.