refs #476
Fixed a few problems with holding down the up/down button on the spinbuttons which would cause multiple programs to be queued up, which at the completion of each, update the widget to the value the novatech contains, affectively preventing you from stepping through the value at a rate faster than it could program the novatech!

Also cleaned up the novatech  __init__ code a bit (changes based on the Pulseblaster init code)!