device_base_class.py/tab_frame.ui Started too build in the remote value check code.
tab_base_classes.py
Added in smart programming
rewrote timeout code

This is a commit before an even more significant rewrite of the timeout adding code. I've realised there will be bugs with this given the implementation of various modes, which could potentially lead to the buildup of many statefunctions due to the fact that the timeout_add function can run in all states, but the statefunction you want running every x seconds may stay and build up in the queue if it can only run in manual_mode and we are in another mode.