Modified the Tab text colour updating code to be more robust
We now avoid having multiple timers running continusously in competition with each other to update the tab text colour. There is a single timer that always runs  when the tab text colour needs updating, and once the tab is available to update, it will update it to the last colour that was specified.