removed duplicate initialisation functions from all tabs.
There used to be two initialisation functions, 'init' and 'initialise' in the worker process.
This was because init was called outside of the state machine, meaning errors were not raised properly.
I have now moved init to be called as a state-function (see changeset:38a49923ef69) and merged all initialise functions into the init method for all device workers.