Modified worker initialisation to always run the worker's init method first
We now manually insert the state that will run Tab._initialise_worker at the start of the queue. Any states queued up before this are thus bumped further down the queue.

Note that while the mainloop is started immediatedly in Tab.__init__ it cannot get an item out of the queue until control is returned to the Qt Event loop and the Qt Event is processed to check for items in the queue (which doesn't happen until after the worker processes are created by DeviceTab.__init__ )